backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
redis[hiredis]>=5.0.0
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
python-dotenv==1.0.0
alembic==1.12.1
pytest-asyncio==0.23.2
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
slowapi==0.1.8
prometheus-client==0.17.1
tenacity==8.2.3
psutil==5.9.0
aioredis>=2.0.0
pytest-mock==3.12.0
google-api-python-client>=2.0.0